Add files via upload
To translate the visual gestures into audible speech, the system employs the pyttsx3 module, converting text labels that correspond to recognized gestures into speech. This integration ensures that each gesture is not only identified but also labeled and stored, creating a robust dataset for both the training and testing phases. The machine learning model at the heart of this system is crafted using Teachable Machine, a web tool from Google Creative Lab that leverages TensorFlow.js for training purposes. This trained model is then seamlessly integrated into the application, enabling the real-time operation of the sign language interpreter.